+++ src/Order/OrderService.php
@@ -14,9 +14,11 @@ use GuzzleHttp\RequestOptions;
 use Psr\Http\Message\StreamInterface;
 use Wizaplace\SDK\AbstractService;
 use Wizaplace\SDK\Authentication\AuthenticationRequired;
+use Wizaplace\SDK\Exception\AccessDenied;
 use Wizaplace\SDK\Exception\NotFound;
 use Wizaplace\SDK\Exception\OrderNotCancellable;
 use Wizaplace\SDK\Exception\SomeParametersAreInvalid;
+use Wizaplace\SDK\Exception\UnauthorizedModerationAction;
 use Wizaplace\SDK\Subscription\SubscriptionSummary;
 
 /**
@@ -466,5 +468,31 @@ final class OrderService extends AbstractService
             }
             throw $e;
         }
+    }
+
+    public function dispatchFunds(int $orderId): array
+    {
+        $this->client->mustBeAuthenticated();
+
+        try {
+            $response = $this->client->post('orders/' . $orderId . '/dispatch');
+        } catch (ClientException $e) {
+            if ($e->getResponse()->getStatusCode() === 400) {
+                throw new SomeParametersAreInvalid("The Dispatch could not be processed. Check out the response payload to identify the issue.", 400, $e);
+            }
+            if ($e->getResponse()->getStatusCode() === 401) {
+                throw new UnauthorizedModerationAction("Unauthorized access.");
+            }
+            if ($e->getResponse()->getStatusCode() === 403) {
+                throw new AccessDenied("Insufficient permission.");
+            }
+            if ($e->getResponse()->getStatusCode() === 404) {
+                throw new NotFound("Order #{$orderId} not found", $e);
+            }
+
+            throw $e;
+        }
+
+        return $response;
     }
 }